extract script to generate QR code
