+++ example-payload/slr-example-2.json
@@ -1,0 +1,59 @@
+{
+  "p": {
+    "fnt": "VON<TESTBETRIEB",
+    "gnt": "THEODOR<GUENTHER",
+    "fn": "von Testbertieb",
+    "gn": "Theodor Günther",
+    "dob": "1972-12-12"
+  },
+  "mo": {
+    "id": "01DE/4967998/DXSGWLWL40SU8ZFKIYIBK39A3#S",
+    "ds": "2023-01-15",
+    "dr": "2023-01-16",
+    "mor": {
+      "2339-0": 0,
+      "59032-3": 0,
+      "35254-2": 0,
+      "1747-5": 2,
+      "42719-5": 2,
+      "1971-1": 2,
+      "1968-7": 2,
+      "1920-8": 0,
+      "1742-6": 0,
+      "6768-6": 0,
+      "2532-0": 0,
+      "2157-6": 0,
+      "1798-8": 0,
+      "2093-3": 0,
+      "2085-9": 0,
+      "43396-1": 0,
+      "2089-1": 0,
+      "3043-7": 0,
+      "3084-1": 0,
+      "3092-4": 0,
+      "3095-7": 0,
+      "2160-0": 0,
+      "16688-4": 2,
+      "62238-1": 2,
+      "2951-2": 2,
+      "19123-9": 2,
+      "LP65359-9": 0,
+      "46099-8": 2,
+      "LP14912-7": 2,
+      "2498-4": 2,
+      "2276-4": 2,
+      "3034-6": 2,
+      "2502-3": 0,
+      "3016-3": 0,
+      "83126-3": 2,
+      "3024-7": 2,
+      "2143-6": 0,
+      "2692-2": 2,
+      "21198-7": 2,
+      "10835-7": 0
+    }
+  },
+  "di": "2022-02-03",
+  "ci": "svlr:1:ub:DE/JBI32302/4Ef57666/22",
+  "ver": "1.0.0"
+}
